requirements.txt
Flask==2.3.3
dnspython==2.4.2